refactor: destructured projectsPage client side
